Doc: IR: Make property nodes anchored sections
Allow <texi-node> objects to render as TeXinfo sections only, without
an associated actual TeXinfo node. Set an anchor, however, so that
these objects can still be cross-referenced.

Introduce a <texi-node> method node-title to create the section title
string. Overriding this allows to display a title different from the
node name.

Make use of both features for property documentation nodes. This gets
rid of the rather redundant node menu and allows to add the associated
value type for each property to the node title and to use @code type
for the symbol itself.